requirements.txt
Django==1.10
django-bootstrap3==7.1.0
django-libsass==0.7
django-allauth==0.27.0
django-compressor==2.1
libsass==0.11.1
psycopg2==2.6.2
dj-database-url==0.4.1